chatbot/requirements.txt
google-generativeai==0.8.4
gtts==2.5.4
pygame==2.5.2
SpeechRecognition==3.10.1
langdetect==1.0.9
pydub==0.25.1
flask==3.0.2
python-dotenv==1.0.1
flask-cors